P2-7: ProgressCursor.lastDocProcessed int→long, start at 0
Change ProgressCursor.lastDocProcessed from int to long to support
shards with >2B documents. Propagate the type change through:
- WorkItemCursor.progressCheckpointNum (int→long)
- WorkItem.startingDocId (Integer→Long)
- MigrationPipeline.migrateShard offset tracking (int→long)
- DocumentSource/LuceneSnapshotSource/SyntheticDocumentSource (int→long)
- PipelineDocumentsRunner startingOffset (int→long)

Change initial work item startingDocId from Integer.MIN_VALUE to 0,
removing the negative sentinel value.

+++ SnapshotReader/src/main/java/org/opensearch/migrations/bulkload/pipeline/adapter/LuceneSnapshotSource.java
@@ -110,7 +110,7 @@ public class LuceneSnapshotSource implements DocumentSource {
     }
 
     @Override
-    public Flux<DocumentChange> readDocuments(ShardId shardId, int startingDocOffset) {
+    public Flux<DocumentChange> readDocuments(ShardId shardId, long startingDocOffset) {
         var entry = resolveShardEntry(shardId, shardEntryCache);
         if (entry == null) {
             return Flux.error(new IllegalArgumentException("Shard not found: " + shardId));
@@ -132,7 +132,7 @@ public class LuceneSnapshotSource implements DocumentSource {
     }
 
     private Flux<DocumentChange> readRegularDocuments(
-        SnapshotExtractor.ShardEntry entry, ShardId shardId, int startingDocOffset
+        SnapshotExtractor.ShardEntry entry, ShardId shardId, long startingDocOffset
     ) {
         log.info("Reading documents from {} starting at offset {}", shardId, startingDocOffset);
         return extractor.readDocuments(entry, workDir)
